+++ wavesexchange_repos/src/lib.rs
@@ -1,4 +1,1 @@
-#[macro_use]
-extern crate async_trait;
-
 pub mod circuit_breaker;
